Bug #13266 FRV/AISD: Uniproxy does'nt log incoming data binary LF -> CR/LF. Remove warnings and bump version

+++ src/applutil.cpp
@@ -720,7 +720,7 @@ std::string readfile( const std::string &_filename )
    int length;
    std::ifstream ifs( _filename );
    ifs.seekg( 0, std::ios::end );
-   length = ifs.tellg();
+   length = static_cast<int>(ifs.tellg());
    ifs.seekg( 0, std::ios::beg );
    if ( length > 0 )
    {
